third_party/blink/web_tests/http/tests/eventsource/eventsource-url-attribute-expected.txt

Only .url should work, previously supported .URL should not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS source.url is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS source.URL is undefined.
PASS successfullyParsed is true

TEST COMPLETE
